backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-core
faiss-cpu
sentence-transformers
pypdf
python-dotenv
groq
sqlalchemy
beautifulsoup4
requests
tiktoken